+++ internal/application/dto/user.go
@@ -5,5 +5,5 @@ import "github.com/google/uuid"
 type User struct {
 	ID      uuid.UUID `json:"id"`
 	Name    string    `json:"name"`
-	Balance float64   `json:"balance"`
+	Balance float32   `json:"balance"`
 }
